small fixes

+++ api/server.js
@@ -53,9 +53,15 @@ app.post('/auth/login', (req, res)=>{
 app.post('/auth/register', multipartyMiddleWare ,(req, res)=>{
     const username = req.body.username
     let password = req.body.password
-    const image = req.files.image.path.split("uploads\\").pop()
+    let image;
+    if(req.files && Object.keys(req.files).length != 0) {
+        image = req.files.image.path.split("uploads\\").pop()
+    } else {
+        image = null
+    }
     const date = new Date()
     const mysqlDate = `${date.getFullYear()}-${date.getMonth()}-${date.getDate()}`
+    console.log("ja hier komt ie");
     
 
     password = crypto.createHmac("sha256", "password")
@@ -68,11 +74,15 @@ app.post('/auth/register', multipartyMiddleWare ,(req, res)=>{
     [username, password, image, mysqlDate, mysqlDate],
     (err, result)=>{
         if(err){
-            fs.unlinkSync(`./uploads/${image}`)
+            if(image) {
+                fs.unlinkSync(`./uploads/${image}`)
+            }
             return res.send(err)
         }
-        if(!fs.existsSync(`./uploads/${hashedUsername}`)) fs.mkdirSync(`./uploads/${hashedUsername}`);
-        fs.renameSync(`./uploads/${image}`, `./uploads/${hashedUsername}/${image}`)
+        if(image) {
+            if(!fs.existsSync(`./uploads/${hashedUsername}`)) fs.mkdirSync(`./uploads/${hashedUsername}`);
+            fs.renameSync(`./uploads/${image}`, `./uploads/${hashedUsername}/${image}`)
+        }
         return res.send(result)
     })
 })
